Extract GameScene's Observer extension to new file
